requirements.txt
commonmark==0.9.1
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-frontmatter==0.5.0
PyYAML==5.3
six==1.13.0
